postgresql: Rework how we setup pg_hba entries required by crowbar
This was somehow magic until now, as it depended on the number of
entries to do the right thing, and was always assuming nothing else was
touching the entries.

Now we properly track the ones that are automatically added by the
recipe, so we can update them easily when needed.