neutron: Add ability to specify agent_boot_time
With L2pop, when the first port on a node gets active, or when the agent
is still starting (uptime of agent  < agent_boot_time), then a full sync
of the L2pop for that network is done.

The sync therefore does not happen if there are more than one port on
the node (like dhcp + router) and if the agent gets the port
notification after start+agent_boot_time.

The first condition can occur fairly easily, and the second definitely
happens when the cloud is large and/or on boot (because the l3 agent
and the dhcp agent are start after the ovs agent, and sometimes the
addition of the ports can hence happen long after ovs agent has
started).

Closes https://github.com/sap-oc/crowbar-openstack/issues/36